+++ doc/whatsnew.rst
@@ -39,6 +39,10 @@ v0.3.1.dev
 
 - Add example in gallery on frequency band selection using class distinctiveness measure. :pr:`219` by :user:`MSYamamoto`
 
+- Add :func:`pyriemann.utils.covariance.covariance_mest` supporting three robust M-estimators (Huber, Student-t and Tyler)
+  and available for all covariance based functions and classes; and add an example on robust covariance estimation for corrupted data.
+  Add also :func:`pyriemann.utils.distance.distance_mahalanobis` between between vectors and a Gaussian distribution. :pr:`223` by :user:`qbarthelemy`
+
 v0.3 (July 2022)
 ----------------